requirements.txt
Django==4.0.5
flake8==4.0.1
djangorestframework==3.13.1